Move things from now-detached lib.rs to mod.rs
